refactor: moved VacancyDetail to wigets folder
