+++ src/widgets/vacancy-detail/index.ts
