+++ src/widgets/vacancy-detail/ui/VacancyDetail.tsx
