+++ src/widgets/vacancy-detail/ui/VacancyDetail.types.ts
