+++ src/widgets/vacancy-detail/ui/components/VacancyMainContent/VacancyMainContent.tsx
