+++ src/widgets/vacancy-detail/ui/components/VacancySummaryCard/VacancySummaryCard.types.ts
