+++ src/widgets/vacancy-detail/ui/components/index.ts
